Small fixes after meeting with mentor:
1. Renamed artifact fo be lowercase (Linkedin -> linkedin). According to maven recommendations.
2. In the previous commit, Oksana changes groupId. So I decided to modify folders structure correspondingly. It will be as follows: src/main/java/com/danit/fs12
3. Removed httpStatus from the body of ResponseEntity, because it is already there in the headers of ResponseEntity.
