+++ src/frontend/.gitignore
@@ -2,6 +2,7 @@
 
 # dependencies
 /node_modules
+/node
 /.pnp
 .pnp.js
