+++ src/main/java/com/danit/fs12/dto/UserDtoReq.java
@@ -1,4 +1,4 @@
-package com.linkedin.app.dto;
+package com.danit.fs12.dto;
 
 import lombok.Data;
