+++ src/main/java/com/danit/fs12/exception/NoSuchUserException.java
@@ -1,4 +1,4 @@
-package com.linkedin.app.exception;
+package com.danit.fs12.exception;
 
 public class NoSuchUserException extends RuntimeException {
