+++ src/main/java/com/danit/fs12/service/UserServiceImp.java
@@ -1,0 +1,27 @@
+package com.danit.fs12.service;
+
+import com.danit.fs12.entity.User;
+import com.danit.fs12.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImp implements UserService {
+
+  private final UserRepository userRepository;
+
+  @Override
+  public User getActiveUser(long id) {
+
+    //temporary to test
+    return new User(1,
+        "Gianluigi",
+        "Donnarumma",
+        "gian_donna@gmail.com",
+        "393290233725",
+        22,
+        "gianni",
+        "secret");
+  }
+}
