+++ src/main/java/com/danit/fs12/utils/AuditConfig.java
@@ -1,4 +1,4 @@
-package com.linkedin.app.utils;
+package com.danit.fs12.utils;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
