--- src/main/java/com/linkedin/app/service/UserService.java
@@ -1,10 +1,0 @@
-package com.linkedin.app.service;
-
-import com.linkedin.app.entity.User;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface UserService {
-
-  public User getActiveUser(long id);
-}
